fixed delete a game & try to add the image to the backOffice

+++ src/main/java/pfr/backgamesloc/games/services/GameService.java
@@ -3,13 +3,19 @@ package pfr.backgamesloc.games.services;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 import pfr.backgamesloc.customers.DAL.CustomerRepository;
 import pfr.backgamesloc.customers.DAL.entities.Customer;
 import pfr.backgamesloc.games.DAL.GameRepository;
 import pfr.backgamesloc.games.DAL.entities.Game;
+import pfr.backgamesloc.games.controllers.DTO.GameEditDTO;
 
+import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -33,8 +39,12 @@ public class GameService {
         return this.gameRepository.findAllByOrderByGameIdAsc();
     }
 
-    public Game createANewGame(Game game) {
-        return this.gameRepository.save(game);
+    public Game createANewGame(GameEditDTO gameEditDTO) throws IOException {
+        MultipartFile file = gameEditDTO.getImage();
+        byte[] bytes = file.getBytes();
+        System.out.println(file.getContentType());
+
+        return this.gameRepository.save(new Game());
     }
 
     public List<Game> addGameToFavorites(Integer customerId, Integer gameId) {
@@ -50,12 +60,19 @@ public class GameService {
         return customerRepository.save(customer).getFavoriteGames();
     }
 
-    public Game editGameById(Integer gameId, Game game) {
+    public void editGameById(Integer gameId, Game game) {
         game.setGameId(gameId);
-        return this.createANewGame(game);
+//        return this.createANewGame(game);
     }
 
-    public void deleteById(Integer gameId) {
-        this.gameRepository.deleteById(gameId);
+    public boolean deleteById(Integer gameId) {
+        Optional<Game> gameOptional = this.gameRepository.findById(gameId);
+
+        if (gameOptional.isPresent()) {
+            this.gameRepository.deleteById(gameId);
+            return true;
+        } else {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Game cannot be deleted");
+        }
     }
 }
